requirements.txt
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
click==8.3.3
exceptiongroup==1.3.1
fastapi==0.136.1
greenlet==3.5.0
h11==0.16.0
idna==3.13
pydantic==2.13.3
pydantic_core==2.46.3
pydentic==0.0.1.dev3
python-dotenv==1.2.2
python-stdnum==2.2
SQLAlchemy==2.0.49
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.46.0